add translation for all relevant instructions
I have changed the approach yet again. We now only operate on
unoptimized modules, which feature calls to the @llvm.dbg.declare
intrinsic. This intrinsic lets us find the stack-local value that
represents variables in order to permute based on the instance.

This approach lets us also indicate the location and name of source
variables, which is something to consider for the future.